+++ app/src/main/java/org/simple/clinic/patient/PatientProfile.kt
@@ -1,6 +1,8 @@
 package org.simple.clinic.patient
 
 import android.os.Parcelable
+import androidx.room.Embedded
+import androidx.room.Relation
 import kotlinx.parcelize.Parcelize
 import org.simple.clinic.patient.businessid.BusinessId
 import org.simple.clinic.patient.businessid.Identifier.IdentifierType.IndiaNationalHealthId
@@ -8,9 +10,25 @@ import java.util.UUID
 
 @Parcelize
 data class PatientProfile(
+    @Embedded
     val patient: Patient,
+
+    @Relation(
+        parentColumn = "addressUuid",
+        entityColumn = "uuid"
+    )
     val address: PatientAddress,
+
+    @Relation(
+        parentColumn = "uuid",
+        entityColumn = "patientUuid"
+    )
     val phoneNumbers: List<PatientPhoneNumber>,
+
+    @Relation(
+        parentColumn = "uuid",
+        entityColumn = "patientUuid"
+    )
     val businessIds: List<BusinessId>
 ) : Parcelable {
   val patientUuid: UUID